Fecal_Coliform/E_coli replace '#' outside of replace_unit_by_dict()
replace_unit_by_dict() is based on the entire string, since '#' is hard to deal with it's replaced the same way regardless of where in the unit string it is. There is a current TODO: to determine why doing string replacements before dict replacement (generally preferable as it standardizes units first) were problematic.